Move all random operators into one file
