+++ alpha/ruby/lib/plan_out.rb
@@ -1,11 +1,4 @@
-require_relative 'plan_out/operator'
-require_relative 'plan_out/op_simple'
-require_relative 'plan_out/op_random'
-require_relative 'plan_out/random_float'
-require_relative 'plan_out/random_integer'
-require_relative 'plan_out/bernoulli_trial'
-require_relative 'plan_out/weighted_choice'
-require_relative 'plan_out/uniform_choice'
+require_relative 'plan_out/random_operators'
 require_relative 'plan_out/assignment'
 require_relative 'plan_out/experiment'
 require_relative 'plan_out/simple_experiment'
