--- alpha/ruby/lib/plan_out/bernoulli_trial.rb
@@ -1,11 +1,0 @@
-require_relative 'op_random'
-
-module PlanOut
-  class BernoulliTrial < OpRandom
-    def simple_execute
-      p = @parameters[:p]
-      rand_val = get_uniform(0.0, 1.0)
-      (rand_val <= p) ? 1 : 0
-    end
-  end
-end
