+++ alpha/ruby/lib/plan_out/version.rb
@@ -1,3 +1,3 @@
 module PlanOut
-  VERSION = '0.1.0'
+  VERSION = '0.1.1'
 end
